check if subscription ticker exists as key

+++ app.py
@@ -41,7 +41,7 @@ async def subscribe(channel):
 
     deployment_name = channel.lower() + "-worker"
 
-    if deployment_name not in socket.ws_client.subscriptions["ticker"]:
+    if "ticker" in socket.ws_client.subscriptions.keys() and deployment_name not in socket.ws_client.subscriptions["ticker"]:
         deployment = kubernetesService.create_deployment_object(channel.lower(), IMAGE + ":" + VERSION, PORT,
                                                                 deployment_name, channel)
         kubernetesService.create_deployment(deployment, deployment_name)
@@ -57,7 +57,7 @@ async def unsubscribe(channel):
 
     deployment_name = channel.lower() + "-worker"
 
-    if deployment_name in socket.ws_client.subscriptions["ticker"]:
+    if "ticker" in socket.ws_client.subscriptions.keys() and deployment_name in socket.ws_client.subscriptions["ticker"]:
         kubernetesService.delete_deployment(deployment_name)
     else:
         return {"message": "not previously subscribed to channel: " + channel}
